backend/app_archive/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
aiohttp==3.9.1
asgiref==3.7.2
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
groq==0.13.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hypercorn==0.15.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-json-logger==2.0.7
sniffio==1.3.1
typing_extensions==4.12.2
Werkzeug==3.1.3